+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add replication destination prefix support, [PR-94](https://github.com/reductstore/reduct-rs/pull/94)
+- Add replication compression setting support, [PR-96](https://github.com/reductstore/reduct-rs/pull/96)
 
 ## 1.20.0 - 2026-06-17